# parqueadero-api-openapi.yaml
openapi: 3.0.3
info:
  title: API Gestión Parqueadero
  description: API para la gestión de entradas, salidas, pagos y espacios de un parqueadero.
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Servidor Local
tags:
  - name: Auth
    description: Operaciones de Autenticación
  - name: Usuarios
    description: Gestión de Usuarios (Admin)
  - name: Parqueadero
    description: Operaciones del Parqueadero
paths:
  /api/auth/register:
    post:
      summary: Registrar nuevo usuario
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/login:
    post:
      summary: Iniciar sesión
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/eliminar/{username}:
    delete:
      summary: Eliminar un usuario por username
      tags: [Auth]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Nombre de usuario a eliminar
      responses:
        "204":
          description: Usuario eliminado exitosamente
        "404":
          description: Usuario no encontrado
  /api/usuarios:
    get:
      summary: Lista todos los usuarios y sus roles (Solo para Admin)
      tags: [Usuarios]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de usuarios y sus roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UsuarioDTO"
  /api/usuarios/{username}/rol:
    put:
      summary: Cambia el rol de un usuario (Solo para Admin)
      tags: [Usuarios]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: El nombre de usuario a modificar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CambiarRolRequest"
      responses:
        "200":
          description: Rol actualizado correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioDTO"
        "400":
          description: Rol inválido o datos incorrectos
        "404":
          description: Usuario no encontrado
  /api/parqueadero/entrada:
    post:
      summary: Registrar entrada de vehículo
      tags: [Parqueadero]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntradaRequest"
      responses:
        "200":
          description: Entrada registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDTO"
  /api/parqueadero/salida:
    post:
      summary: Registrar salida de vehículo y calcular pago
      tags: [Parqueadero]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalidaRequest"
      responses:
        "200":
          description: Salida procesada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagoResponse"
  /api/parqueadero/tickets/{codigo}:
    get:
      summary: Obtener ticket por código
      tags: [Parqueadero]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: codigo
          schema:
            type: string
          required: true
          description: Código del ticket (ej. TKT-12345)
      responses:
        "200":
          description: Detalle del ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDTO"
  /api/parqueadero/tickets/activos:
    get:
      summary: Listar tickets activos
      tags: [Parqueadero]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de tickets activos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketDTO"
  /api/parqueadero/espacios:
    get:
      summary: Listar todos los espacios
      tags: [Parqueadero]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de espacios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EspacioDTO"
  /api/parqueadero/espacios/agregar:
    post:
      summary: Agregar nuevos espacios
      tags: [Parqueadero]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgregarEspaciosRequest"
      responses:
        "200":
          description: Espacios creados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EspacioDTO"
  /api/parqueadero/espacios/eliminar:
    delete:
      summary: Eliminar espacios de parqueadero disponibles
      tags: [Parqueadero]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EliminarEspaciosRequest"
      responses:
        "200":
          description: Espacios eliminados exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EspacioDTO"
  /api/parqueadero/estadisticas:
    get:
      summary: Obtener estadísticas del parqueadero
      tags: [Parqueadero]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Estadísticas actuales
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstadisticasDTO"
  /api/parqueadero/tarifas:
    get:
      summary: Lista las tarifas globales configuradas
      tags: [Parqueadero]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de tarifas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TarifaDTO"
    post:
      summary: Crear o actualizar una tarifa global
      description: Busca una tarifa por su 'tipoTarifa'. Si existe, actualiza su 'valor'. Si no existe, crea una nueva tarifa.
      tags: [Parqueadero]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TarifaDTO"
      responses:
        "200":
          description: Tarifa guardada o actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TarifaDTO"
  /api/parqueadero/tarifas/{tipoTarifa}:
    delete:
      summary: Elimina una tarifa global (Admin)
      tags: [Parqueadero]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: tipoTarifa
          schema:
            type: string
          required: true
          description: Tipo de tarifa a eliminar
      responses:
        "204":
          description: Tarifa eliminada exitosamente

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CambiarRolRequest:
      type: object
      properties:
        newRole:
          type: string
          description: El nuevo rol para el usuario.
          enum: [ADMIN, OPERADOR, USER]
    UsuarioDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
    AuthRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
    EntradaRequest:
      type: object
      properties:
        placa:
          type: string
        tipoVehiculo:
          type: string
          enum: [CARRO, MOTO, BICICLETA]
        tipoTarifa:
          type: string
          enum: [POR_MINUTO, POR_HORA, POR_DIA, POR_MES, FRACCION]
    SalidaRequest:
      type: object
      properties:
        codigoTicket:
          type: string
        placa:
          type: string
        observaciones:
          type: string
    AgregarEspaciosRequest:
      type: object
      properties:
        tipoVehiculo:
          type: string
          enum: [CARRO, MOTO, BICICLETA]
        cantidad:
          type: integer
        tarifaBase:
          type: number
    EliminarEspaciosRequest:
      type: object
      properties:
        tipoVehiculo:
          type: string
          enum: [CARRO, MOTO, BICICLETA]
        cantidad:
          type: integer
    TicketDTO:
      type: object
      properties:
        codigo:
          type: string
        placa:
          type: string
        tipoVehiculo:
          type: string
        codigoEspacio:
          type: string
        horaEntrada:
          type: string
          format: date-time
        tipoTarifa:
          type: string
        estado:
          type: string
    EspacioDTO:
      type: object
      properties:
        codigo:
          type: string
        tipoVehiculo:
          type: string
        estado:
          type: string
        tarifaBase:
          type: number
    PagoResponse:
      type: object
      properties:
        codigoTicket:
          type: string
        horaEntrada:
          type: string
          format: date-time
        horaSalida:
          type: string
          format: date-time
        duracionHoras:
          type: integer
        duracionMinutos:
          type: integer
        valorBase:
          type: number
        valorAdicional:
          type: number
        descuento:
          type: number
        valorTotal:
          type: number
        mensaje:
          type: string
    EstadisticasDTO:
      type: object
      properties:
        vehiculosActivos:
          type: integer
        espaciosDisponibles:
          type: integer
        espaciosOcupados:
          type: integer
        ingresosHoy:
          type: number
        ingresosMes:
          type: number
        ticketsHoy:
          type: integer
        ticketsMes:
          type: integer
    TarifaDTO:
      type: object
      properties:
        id:
          type: integer
        tipoTarifa:
          type: string
          enum: [POR_MINUTO, POR_HORA, POR_DIA, POR_MES, FRACCION]
        valor:
          type: number