# src/main/java/com/parqueadero/controller/ParqueaderoController.java
package com.parqueadero.controller;

import com.parqueadero.dto.*;
import com.parqueadero.dto.TarifaDTO;
import com.parqueadero.service.ParqueaderoService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/parqueadero")
@RequiredArgsConstructor
public class ParqueaderoController {

    private final ParqueaderoService parqueaderoService;

    @Operation(summary = "Registra la entrada de un vehículo",
               description = "Crea un nuevo ticket para un vehículo que ingresa al parqueadero. Si el vehículo no existe, se crea.",
               responses = {
                   @ApiResponse(responseCode = "200", description = "Entrada registrada exitosamente"),
                   @ApiResponse(responseCode = "400", description = "Datos de entrada inválidos o vehículo ya estacionado", content = @Content),
                   @ApiResponse(responseCode = "404", description = "No hay espacios disponibles para el tipo de vehículo", content = @Content)
               })
    @PostMapping("/entrada")
    public ResponseEntity<TicketDTO> registrarEntrada(@RequestBody EntradaRequest request) {
        TicketDTO ticket = parqueaderoService.registrarEntrada(request);
        return ResponseEntity.ok(ticket);
    }

    @Operation(summary = "Registra la salida de un vehículo y calcula el pago")
    @PostMapping("/salida")
    public ResponseEntity<PagoResponse> registrarSalida(@RequestBody SalidaRequest request) {
        PagoResponse pago = parqueaderoService.registrarSalida(request);
        return ResponseEntity.ok(pago);
    }

    @GetMapping("/tickets/{codigo}")
    public ResponseEntity<TicketDTO> obtenerTicket(@PathVariable String codigo) {
        return ResponseEntity.ok(parqueaderoService.obtenerTicket(codigo));
    }

    @GetMapping("/tickets/activos")
    public ResponseEntity<List<TicketDTO>> listarTicketsActivos() {
        return ResponseEntity.ok(parqueaderoService.listarTicketsActivos());
    }

    @Operation(summary = "Lista las tarifas globales configuradas",
               description = "Devuelve las tarifas globales como POR_MINUTO, POR_DIA, etc. Las tarifas por hora o fracción son específicas de cada espacio y se pueden consultar en el detalle de cada uno a través del endpoint de /espacios.")
    @GetMapping("/tarifas")
    public ResponseEntity<List<TarifaDTO>> listarTarifas() {
        return ResponseEntity.ok(parqueaderoService.listarTarifas());
    }

    @Operation(summary = "Crea o actualiza una tarifa global (Admin)")
    @PostMapping("/tarifas")
    public ResponseEntity<TarifaDTO> guardarTarifa(@RequestBody TarifaDTO request) {
        return ResponseEntity.ok(parqueaderoService.guardarTarifa(request));
    }

    @GetMapping("/espacios")
    public ResponseEntity<List<EspacioDTO>> listarEspacios() {
        return ResponseEntity.ok(parqueaderoService.listarEspacios());
    }

    @GetMapping("/estadisticas")
    public ResponseEntity<EstadisticasDTO> obtenerEstadisticas() {
        return ResponseEntity.ok(parqueaderoService.obtenerEstadisticas());
    }

    @Operation(summary = "Agrega nuevos espacios de parqueadero (Admin)")
    @PostMapping("/espacios/agregar")
    public ResponseEntity<List<EspacioDTO>> agregarEspacios(@RequestBody AgregarEspaciosRequest request) {
        return ResponseEntity.ok(parqueaderoService.agregarEspacios(request));
    }

    @Operation(summary = "Elimina espacios de parqueadero disponibles (Admin)")
    @DeleteMapping("/espacios/eliminar")
    public ResponseEntity<List<EspacioDTO>> eliminarEspacios(@RequestBody EliminarEspaciosRequest request) {
        return ResponseEntity.ok(parqueaderoService.eliminarEspacios(request));
    }

    @GetMapping("/historial")
    public ResponseEntity<List<HistorialDTO>> listarHistorial() {
        return ResponseEntity.ok(parqueaderoService.listarHistorial());
    }

    @Operation(summary = "Elimina una tarifa global (Admin)")
    @DeleteMapping("/tarifas/{tipoVehiculo}/{tipoTarifa}")
    public ResponseEntity<Void> eliminarTarifa(@PathVariable String tipoVehiculo, @PathVariable String tipoTarifa) {
        parqueaderoService.eliminarTarifa(tipoVehiculo, tipoTarifa);
        return ResponseEntity.noContent().build();
    }
}